requirements.txt
amqp==5.0.2
appdirs==1.4.4
argcomplete==1.12.0
astroid==2.4.2
attrs==20.3.0
Beaker==1.10.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
black==20.8b1
blivet==3.3.0
blivet-gui==2.2.1
Brlapi==0.8.0
celery==5.0.5
cffi==1.14.1
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.0
cupshelpers==1.0
dasbus==1.4
dbus-python==1.2.16
decorator==4.4.2
distro==1.5.0
Flask==1.1.2
fros==1.1
gpg==1.14.0
humanize==0.5.1
idna==2.10
iniconfig==1.1.1
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==5.0.2
langtable==0.0.53
lazy-object-proxy==0.0.0
libcomps==0.1.15
lxml==4.5.1
Mako==1.1.3.dev0
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
nftables==0.1
olefile==0.46
packaging==20.8
Paste==3.4.0
pathspec==0.8.1
pexpect==4.8.0
pid==2.2.3
Pillow==8.0.1
pluggy==0.13.1
ply==3.11
productmd==1.28
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pwquality==1.4.2
py==1.10.0
pycairo==1.18.2
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.43.0.5
pyenchant==3.1.1
PyGObject==3.38.0
pykickstart==3.29
pylint==2.6.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyparted==3.11.5
PySocks==1.7.1
pytest==6.2.1
python-augeas==0.5.0
python-dateutil==2.8.1
python-meh==0.50
pytz==2020.1
pyudev==0.22.0
pyxdg==0.26
redis==3.5.3
regex==2020.11.13
requests==2.24.0
requests-file==1.4.3
requests-ftp==0.3.1
rpm==4.16.0
rq==1.7.0
selinux==3.1
sepolicy==3.1
setools==4.3.0
simpleline==1.7
six==1.15.0
slip==0.6.4
slip.dbus==0.6.4
sos==4.0
soupsieve==2.0.1
systemd-python==234
Tempita==0.5.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.8
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
